Add two convenience functions here to manual convert from the prop tree to a python dictionary and visa versa.
Note: Converting from [dict to] json to the
property tree is not supported because of the
memory allocation/managment/leak/referencing
deleted memory issues it invokes.  Generating a
json string from the property /is/ supported and
doesn't need an external function.

The external function is provided for completeness
and who knows if it might find a use case in the
future.